Restructuring (again) of Ion type
The Ion struct now contains no redundant information: All properties general to the species are in the field speciesproperties (including quantum numbers and energies of levels, specification of Einstein A coefficients, specially-defined g-factors, and nonlinear contributions to the Zeeman shift), while properties particular to the Ion instance are in a few other fields. sublevel_structure has been changed to sublevels, and is no longer a dictionary but simply a list of the sublevels present in the Ion instance. To be more Julian, access to relevant physical quantities is now done through functions: ionsublevels(), ionlevels(), quantumnumbers(), landegf(), stark_shift(), zeeman_shift(), energy(), iontransitions(), einsteinA(), and matrix_element(). These functions account for all the details specified in the speciesproperties field such as nonlinear Zeeman shifts and specially-defined g-factors. For instance, zeeman_shift(ion, sublevel, B) determines both whether or not the g-factor is specified in speciesproperties and as such decides whether to calculate it or use the specified value, and also decides whether it needs to add a nonlinear contribution. As another example, energy(ion, sublevel, B=B) returns the energy of the specified sublevel at the specified B-field, also including the Stark shift. This restructuring favors using functions over directly accessing the struct fields. Still to be done is updating hamiltonians.jl to accommodate this and updating the Base package override functions.